requirements.txt
# Flask and Extensions
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1

# Database
SQLAlchemy==1.4.23
alembic==1.7.1

# Forms and Validation
WTForms==2.3.3

# Markdown Processing
Markdown==3.3.4
python-markdown-math==0.8
Pygments==2.10.0

# Security
python-dotenv==0.19.0
pyOpenSSL==20.0.1
cryptography==3.4.7

# Utils
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
gunicorn==20.1.0

# Development
pytest==6.2.5
coverage==5.5
